Add two attributes to support dark/light mode and device type selection
Propose to adding two attributes: color_mode and device_type.
For color_mode, it allows developers to define the default theme color, e,.g. light, dark, or automatically adjust with the system. Can be used for branding, improved visibility for people with poor vision and strong light, and better use of MiniApp in low light environments.
For device_type, it allows developers to specify the type of devices on which the MiniApp can be installed to support cross-platform capabilities of the same app, reducing repeated development. The device types may include mobile phones, tablets, vehicle-mounted devices, wearable devices, TVs, and IoT devices.